PXC-3128: High priority transactions should work with WSREP
Issue: when wsrep is running, high priority transactions weren't able to
kill normal priority transactions. This basically broke the scenario
where a PXC node is also part of a normal replication setup.

Root cause: we intentionally broke this feature in 5.7, in commit
5be95d8ff766. The comment/commit message explains that high priority
transactions aren't how galera works, an dshould be disabled.

Compared to that statement codership fixed HPTs also in 5.7, in commit
f2dd26e10785, by translating the HPT kill logic into wsrep functions.

After merging the codership fix in 5.7, we never reverted our
workaround.

For 8.0, we didn't even port f2dd26e10785, we only ported our disabling
"patch".

Fix:

* Reverting 5be95d8ff766
* Porting f2dd26e10785 to 8.0/galera4
* Removed the assert(0)
* Adding some of the high priority tests to the galera suite to ensure
that they don't break in the future
* This commit also improves the start_transactin_high_prio mtr include
so it works when used together with --debug

Remaining issue: as shown by two TODO comments in the tests, we report a
different error message in some cases compared to basic InnoDB behavior.
This doesn't result in any other behavior change however, we can fix it
later when time permits.
