+++ mysql-test/suite/galera/t/high_prio_trx_2.test
@@ -1,0 +1,58 @@
+# Scenario:
+#  T1=({R(B)})
+#  T2=({R(B), W(B), C}, HIGH_PRIORITY).
+#
+# Outcome: T1 must abort, T2 must commit.
+
+--disable_query_log
+CALL mtr.add_suppression("InnoDB High Priority being used");
+--enable_query_log
+
+--source include/galera_cluster.inc
+
+--source include/count_sessions.inc
+
+CREATE TABLE t1 (c1 INT NOT NULL PRIMARY KEY) ENGINE=InnoDB;
+INSERT INTO t1 VALUES (0);
+
+--connect(con1,localhost,root,,test, $NODE_MYPORT_1)
+--connect(con2,localhost,root,,test, $NODE_MYPORT_1)
+
+--echo
+--echo # On connection 1
+--connection con1
+START TRANSACTION;
+SELECT * FROM t1 WHERE c1=0 FOR UPDATE;
+
+--echo
+--echo # On connection 2
+--connection con2
+--source include/start_transaction_high_prio.inc
+UPDATE t1 SET c1=2 WHERE c1=0;
+COMMIT;
+--disconnect con2
+
+--echo
+--echo # On connection 1
+--connection con1
+# Note: DEADLOCK implies a full ROLLBACK
+--error ER_LOCK_DEADLOCK
+UPDATE t1 SET c1=1 WHERE c1=0;
+--disconnect con1
+
+--connection node_2
+--let $assert_text= 'There is a 2 in t1'
+--let $assert_cond= [SELECT COUNT(*) AS count FROM t1 WHERE t1.c1 = 2, count, 1] = 1
+--source include/assert.inc
+
+--let $assert_text= 'There is no 1 in t1'
+--let $assert_cond= [SELECT COUNT(*) AS count FROM t1 WHERE t1.c1 = 1, count, 1] = 0
+--source include/assert.inc
+
+--let $assert_text= 'There is no 0 in t1'
+--let $assert_cond= [SELECT COUNT(*) AS count FROM t1 WHERE t1.c1 = 0, count, 1] = 0
+--source include/assert.inc
+
+DROP TABLE t1;
+
+--source include/wait_until_count_sessions.inc
